main-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
python-dotenv==1.0.1
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1
jinja2==3.1.3
python-multipart==0.0.9
